sorted files to make them easier to access and simpler to understand
