+++ src/main/java/com/example/group07/project2/apis/UserListApi.java
@@ -1,5 +1,7 @@
-package com.example.group07.project2;
+package com.example.group07.project2.apis;
 
+import com.example.group07.project2.models.UserList;
+import com.example.group07.project2.repos.UserListRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
