+++ src/main/java/com/example/group07/project2/models/User.java
@@ -1,19 +1,22 @@
-package com.example.group07.project2;
+package com.example.group07.project2.models;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
+/**
+ * We are creating a User entity that represents
+ * a user in our database. This is similar to a room database
+ */
 @Entity
-public class UserList {
-
+public class User {
     @Id // this is our primary key
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Integer userId;
 
-    String title;
-    String description;
+    private String username;
+    private String password;
 
     /**
      * Getters and Setters for our private variables
@@ -27,23 +30,19 @@ public class UserList {
         this.userId = id;
     }
 
-    public String getTitle() {
-        return title;
+    public String getUsername() {
+        return username;
     }
 
-    public void setTitle(String title) {
-        this.title = title;
+    public void setUsername(String username) {
+        this.username = username;
     }
 
-    public String getDescription() {
-        return description;
+    public String getPassword() {
+        return password;
     }
 
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public void setList(String list) {
-        this.title=list;
+    public void setPassword(String password) {
+        this.password = password;
     }
 }
