+++ src/main/java/com/example/group07/project2/repos/UserListRepository.java
@@ -1,5 +1,6 @@
-package com.example.group07.project2;
+package com.example.group07.project2.repos;
 
+import com.example.group07.project2.models.UserList;
 import org.springframework.data.repository.CrudRepository;
 
 /**
